Triangles find first

+++ src/main/java/com/playtika/automation/HomeTasks/TriangleTask/Point.java
@@ -4,19 +4,10 @@ package com.playtika.automation.HomeTasks.TriangleTask;
 import java.util.concurrent.ThreadLocalRandom;
 
 public class Point {
-    private int x;
-    private int y;
-
-//    public Point(float x, float y) {
-//        this.x = x;
-//        this.y = y;
-//    }
+    private float x;
+    private float y;
 
     public Point(){
-//        int min = 1;
-//        int max = 2;
-//        this.x = (float)(min + Math.random() * (max - min));
-//        this.y = (float)(min + Math.random() * (max - min));
 
         this.x = ThreadLocalRandom.current().nextInt(1, 5 + 1);
         this.y = ThreadLocalRandom.current().nextInt(2, 5 + 1);
